Trying to minimize changes.
